Move click event to base class
