+++ js/ui/scheduler/tooltip_strategies/desktopTooltipStrategy.js
@@ -5,7 +5,6 @@ import translator from "../../../animation/translator";
 import dragEvents from "../../../events/drag";
 import eventsEngine from "../../../events/core/events_engine";
 import FunctionTemplate from "../../widget/function_template";
-import { extendFromObject } from "../../../core/utils/extend";
 
 const APPOINTMENT_TOOLTIP_WRAPPER_CLASS = "dx-scheduler-appointment-tooltip-wrapper";
 const ALL_DAY_PANEL_APPOINTMENT_CLASS = 'dx-scheduler-all-day-appointment';
@@ -42,6 +41,10 @@ class TooltipSingleAppointmentBehavior extends TooltipBehaviorBase {
     onListItemClick(e) {
         this.scheduler.showAppointmentPopup(e.itemData.data, false, e.itemData.currentData);
     }
+
+    canRaiseClickEvent() {
+        return false;
+    }
 }
 
 class TooltipManyAppointmentsBehavior extends TooltipBehaviorBase {
@@ -55,42 +58,15 @@ class TooltipManyAppointmentsBehavior extends TooltipBehaviorBase {
         }
     }
 
-    onListItemClick(listItemClickArg) {
-        const config = {
-            itemData: listItemClickArg.itemData.data,
-            itemElement: listItemClickArg.itemElement
-        };
-        const showEditAppointmentPopupAction = this.createAppointmentClickAction();
-        showEditAppointmentPopupAction(this.createClickEventArgument(config, listItemClickArg));
-    }
-
-    createAppointmentClickAction() {
-        return this.scheduler._createActionByOption("onAppointmentClick", {
-            afterExecute: e => {
-                const config = e.args[0];
-                config.event.stopPropagation();
-                this.scheduler.fire("showEditAppointmentPopup", { data: config.appointmentData });
-            }
-        });
-    }
-
-    createClickEventArgument(config, listItemClickArg) {
-        const result = extendFromObject(this.scheduler.fire("mapAppointmentFields", config), listItemClickArg, false);
-        return this.trimClickEventArgument(result);
-    }
-
-    trimClickEventArgument(e) {
-        delete e.itemData;
-        delete e.itemIndex;
-        delete e.itemElement;
-        return e;
+    canRaiseClickEvent() {
+        return true;
     }
 
     createFunctionTemplate(template, data, targetData, index) {
         if(this._isEmptyDropDownAppointmentTemplate()) {
             return super.createFunctionTemplate(template, data, targetData, index);
         }
-        return new FunctionTemplate((options) => {
+        return new FunctionTemplate(options => {
             return template.render({
                 model: data,
                 index: index,
@@ -140,7 +116,7 @@ class TooltipManyAppointmentsBehavior extends TooltipBehaviorBase {
             coordinates: coordinates,
             allDay: allDay,
             isFixedContainer: false,
-            callback: (result) => {
+            callback: result => {
                 if(result) {
                     coordinates = result;
                 }
@@ -261,8 +237,7 @@ export class DesktopTooltipStrategy extends TooltipStrategyBase {
         this.behavior.onListItemRendered(e);
     }
 
-    _onListItemClick(e) {
-        super._onListItemClick(e);
-        this.behavior.onListItemClick(e);
+    _canRaiseClickEvent() {
+        return this.behavior.canRaiseClickEvent();
     }
 }
