+++ js/ui/scheduler/tooltip_strategies/tooltipStrategyBase.js
@@ -3,6 +3,7 @@ import dateUtils from "../../../core/utils/date";
 import FunctionTemplate from "../../widget/function_template";
 import $ from "../../../core/renderer";
 import List from "../../list/ui.list.edit";
+import { extendFromObject } from "../../../core/utils/extend";
 
 const TOOLTIP_APPOINTMENT_ITEM = "dx-tooltip-appointment-item",
     TOOLTIP_APPOINTMENT_ITEM_CONTENT = TOOLTIP_APPOINTMENT_ITEM + "-content",
@@ -108,6 +109,42 @@ export class TooltipStrategyBase {
 
     _onListItemClick(e) {
         this.hide();
+        this._canRaiseClickEvent() && this._raiseClickEvent(e);
+    }
+
+    _canRaiseClickEvent() {
+        return true;
+    }
+
+    _raiseClickEvent(e) {
+        const config = {
+            itemData: e.itemData.data,
+            itemElement: e.itemElement
+        };
+        const showEditAppointmentPopupAction = this.createAppointmentClickAction();
+        showEditAppointmentPopupAction(this.createClickEventArgument(config, e));
+    }
+
+    createAppointmentClickAction() {
+        return this.scheduler._createActionByOption("onAppointmentClick", {
+            afterExecute: e => {
+                const config = e.args[0];
+                config.event.stopPropagation();
+                this.scheduler.fire("showEditAppointmentPopup", { data: config.appointmentData });
+            }
+        });
+    }
+
+    createClickEventArgument(config, clickArg) {
+        const result = extendFromObject(this.scheduler.fire("mapAppointmentFields", config), clickArg, false);
+        return this.trimClickEventArgument(result);
+    }
+
+    trimClickEventArgument(e) {
+        delete e.itemData;
+        delete e.itemIndex;
+        delete e.itemElement;
+        return e;
     }
 
     _onDeleteButtonClick() {
